Fix sbpl_recovery
This makes sbpl_recovery run on Kinetic.

I think it's okay to remove the clearNonLethalWindow calls altogether.
If the user wants the costmap to be cleared, they can add the
clear_costmap_recovery before the sbpl_recovery in the
recovery_behaviors list.

Otherwise we would have to steal the clear() function from the
clear_costmap_recovery plugin.

Tested and working.

Fixes #9.